+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update `README.md` to include bundled dependencies url, etc.
 - Remove `vue` from `disableLanguages` because Vue is supported in Prettier 1.15 version.
 - Move `eslint`, `tslint` and `stylelint` integrations to separate file `integrations.ts` (out from `utils.ts`) and `require()` (load) dependency only when it's needed and cache it in-memory.
+- Remove local prettier support.
 
 ## 2.0.0 (August 16, 2019)